requirements-local.txt
pytest==9.0.2
pymysql==1.0.2
matplotlib==3.10.8
beautifulsoup4==4.14.3
pandas==2.3.3
python-dotenv==1.2.1
selenium==4.39.0
SQLAlchemy==2.0.45
yfinance==1.0
loguru==0.7.3
notebook==7.5.3
pytest==9.0.2
pendulum==3.2.0